Replace Part compare method with 'compatible?'
Replace the compare method for Part with a static method that
takes an array of Part objects. Change some tests to reconcile
disagreement between the old compare method and the validity
condition for computers, which were intended to have the same
behavior. The compatible? method just checks if all the
compatibility numbers are the same, allowing parts with the same
name.